Scraping news to choose problem of the day
Script in file ``editorial.py`` scrapes the editorial of the day from La
Republica and tries to guess the problem of the day by couting the number of
times each of the problems from ``problem_list`` appear in the text.

File ``columnista.py`` will call the functions from ``editorial.py`` and if the
editorial deals with a different subject then a random problem will be
selected.

It takes a lot of time for the script to scrape the La Republica site and
extract the problem of the day. It should be better to cache the day's
editorial locally in a file or database for quicker responses every time users
hit F5 to get a new column.